Task_28_final/requirements.txt
pytest
pytest-selenium
termcolor
allure-python-commons
pytest_check
